Homework + exercises Week 1
