+++ week1/possible-qa-exercises/1-extreme-hobbies/extreme-hobbies/src/components/hobby.jsx
@@ -1,0 +1,5 @@
+function Hobby(props) {
+   return <h3 key={props.key}>{props.hobbyName}</h3>
+}
+
+export default Hobby
